[Mail] Moving Mail classes to a new Manager namespace for better organization.
