+++ src/Valkyrja/Application/Env.php
@@ -56,8 +56,8 @@ use Valkyrja\Jwt\Enum\Algorithm;
 use Valkyrja\Jwt\FirebaseJwt;
 use Valkyrja\Log\Contract\Logger;
 use Valkyrja\Log\PsrLogger;
-use Valkyrja\Mail\Contract\Mailer;
-use Valkyrja\Mail\MailgunMailer;
+use Valkyrja\Mail\Manager\Contract\Mailer;
+use Valkyrja\Mail\Manager\MailgunMailer;
 use Valkyrja\Orm\Manager\Contract\Manager;
 use Valkyrja\Orm\Manager\MysqlManager;
 use Valkyrja\Session\Manager\Contract\Session;
