+++ src/Valkyrja/Notification/Manager/Notification.php
@@ -17,8 +17,8 @@ use Override;
 use Valkyrja\Broadcast\Contract\Broadcaster;
 use Valkyrja\Broadcast\Data\Message as BroadcastMessage;
 use Valkyrja\Exception\InvalidArgumentException;
-use Valkyrja\Mail\Contract\Mailer;
 use Valkyrja\Mail\Data\Message as MailMessage;
+use Valkyrja\Mail\Manager\Contract\Mailer;
 use Valkyrja\Notification\Data\Contract\Notify;
 use Valkyrja\Notification\Entity\Contract\NotifiableUser;
 use Valkyrja\Notification\Factory\Contract\Factory;
