+++ tests/Unit/Notification/Provider/ServiceProviderTest.php
@@ -15,7 +15,7 @@ namespace Valkyrja\Tests\Unit\Notification\Provider;
 
 use PHPUnit\Framework\MockObject\Exception;
 use Valkyrja\Broadcast\Contract\Broadcaster;
-use Valkyrja\Mail\Contract\Mailer;
+use Valkyrja\Mail\Manager\Contract\Mailer;
 use Valkyrja\Notification\Factory\ContainerFactory;
 use Valkyrja\Notification\Factory\Contract\Factory;
 use Valkyrja\Notification\Manager\Contract\Notification as Contract;
